Move ModelSearch to separate folder
