+++ app/assets/javascripts/3dbio_viewer/src/webapp/components/model-search/ModelSearch.css
@@ -1,0 +1,11 @@
+.results .models {
+    display: flex;
+}
+
+.results img {
+    width: 11vw;
+}
+
+.model-search .params {
+    display: flex;
+}
