+++ app/assets/javascripts/3dbio_viewer/src/webapp/components/viewer-selector/ViewerSelector.css
@@ -21,15 +21,3 @@
 .selection .selected {
     background-color: #ccc;
 }
-
-.results .models {
-    display: flex;
-}
-
-.results img {
-    width: 11vw;
-}
-
-.model-search .params {
-    display: flex;
-}
